Homework06

+++ HomeWork06/src/main/java/ru/otus/gpbu/pse/homework06/myybooks/repository/AuthorRepository.java
@@ -12,7 +12,7 @@ public interface AuthorRepository {
 
     void update(Author author);
 
-    int deleteById(long id);
+    long deleteById(long id);
 
     List<Author> getAll();
